requirements.txt
PyQt5
PyQtChart
PyYAML
psutil

#windows
wmi
pywin32